005 random

+++ 004 dev/Program.cs
@@ -1,8 +1,8 @@
 ﻿Console.WriteLine("Введите первое число ");
 
-double numberA = Convert.ToInt32(Console.ReadLine());
+double numberA = Convert.ToDouble(Console.ReadLine());
 Console.WriteLine("Введите второе число ");
 
-double numberB = Convert.ToInt32(Console.ReadLine());
+double numberB = Convert.ToDouble(Console.ReadLine());
 double result = numberA / numberB;
 Console.Write($"{numberA}/{numberB}={result}");
